@@ -7,7 +7,10 @@
         "npm": "6.14.8"
     },
     "dependencies": {
-        "@material-ui/icons": "^4.11.3",
+        "@emotion/react": "^11.10.5",
+        "@emotion/styled": "^11.10.5",
+        "@mui/icons-material": "^5.11.0",
+        "@mui/material": "^5.11.1",
         "@testing-library/jest-dom": "^4.2.4",
         "@testing-library/react": "^9.5.0",
         "@testing-library/user-event": "^7.2.1",
@@ -19,19 +22,17 @@
         "jodit-react": "^1.3.24",
         "json-server": "^0.16.1",
         "jsonwebtoken": "^8.5.1",
-        "react": "^16.14.0",
+        "react": "^17.0.0",
         "react-bootstrap": "^1.6.1",
         "react-bootstrap-datetimepicker": "0.0.22",
-        "react-datepicker": "^2.16.0",
+        "react-datepicker": "^4.8.0",
         "react-datetime-picker": "^3.0.2",
-        "react-dom": "^16.14.0",
-        "react-facebook-login": "^4.1.1",
+        "react-dom": "^17.0.0",
         "react-icons": "^4.3.1",
         "react-js-pagination": "^3.0.3",
         "react-jwt": "^1.1.2",
-        "react-load-image": "^0.1.7",
         "react-loading": "^2.0.3",
-        "react-notifications-component": "^2.1.0",
+        "react-notifications-component": "^3.3.3",
         "react-owl-carousel": "^2.3.3",
         "react-redux": "^7.2.0",
         "react-router": "^5.2.0",
