@@ -13,8 +13,8 @@ import "react-datepicker/dist/react-datepicker.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import DatePicker from "react-datepicker";
 import Carousel from "react-bootstrap/Carousel";
-import NavigateNextIcon from "@material-ui/icons/NavigateNext";
-import NavigateBeforeIcon from "@material-ui/icons/NavigateBefore";
+import NavigateNextIcon from '@mui/icons-material/NavigateNext';
+import NavigateBeforeIcon from '@mui/icons-material/NavigateBefore';
 import * as apiUpload from "../../../api/loai_giay";
 import Modal from "react-bootstrap/Modal";
 import Button from "react-bootstrap/Button";
